Project-timeseries_forecasting/pyproject.toml
[tool.poetry]
name = "stat3007-timeseries-forecasting"
version = "0.1.0"
description = ""
authors = ["Lars Moan <lars.ostberg.moan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.4.2"
torch = "2.0.1"
pandas = "^2.2.2"
numpy = "^1.26.4"
matplotlib = "^3.9.0"
wandb = "^0.17.0"
yfinance = "^0.2.40"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
